+++ app/Http/Controllers/CampusController.php
@@ -4,12 +4,13 @@ namespace App\Http\Controllers;
 
 use App\Models\Campus;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class CampusController extends Controller
 {
 
     public function index() {
-        return view('content.campus.table', ['campuses' => Campus::all(),]);
+        return view('content.campus.table', ['campuses' => Auth::user()->campuses,]);
     }
 
     public function create() {
@@ -17,30 +18,44 @@ class CampusController extends Controller
     }
 
     public function store(Request $request) {
-        $campus = Campus::create([
+        $campus = Auth::user()->campuses()->create([
             'name' => $request->input('name'),
             'address' => $request->input('address')
         ]);
-        $campus->save();
+        return redirect()->route('all_campus');
     }
 
     public function view(Campus $campus) {
-        return view('content.campus.view', ['campus', $campus]);
+        if ($campus->user != Auth::user()) { return abort(403); }
+        return view('content.campus.view', ['campus' => $campus]);
     }
 
     public function edit(Campus $campus) {
-        return view('content.campus.update', ['campus', $campus]);
+        if ($campus->user != Auth::user()) { return abort(403); }
+        return view('content.campus.update', ['campus' => $campus]);
     }
 
     public function update(Request $request, Campus $campus) {
+        if ($campus->user != Auth::user()) { return abort(403); }
         $campus->update([
             'name' => $request->input('name'),
             'address' => $request->input('address')
         ]);
         $campus->save();
+        return redirect()->route('all_campus');
     }
 
     public function destroy(Campus $campus) {
+        if ($campus->user != Auth::user()) { return abort(403); }
+        foreach ($campus->buildings as $building) {
+            foreach ($building->switches as $switch) {
+                foreach ($switch->ports as $port) {
+                    $port->delete();
+                }
+                $switch->delete();
+            }
+            $building->delete();
+        }
         $campus->delete();
         return redirect()->route('all_campus');
     }
